+++ src/main/java/com/github/bartimaeusnek/croploadcore/CropLoadCore.java
@@ -1,4 +1,4 @@
-package lokko12.croploadcore;
+package com.github.bartimaeusnek.croploadcore;
 
 import java.io.File;
 import java.security.cert.Certificate;
@@ -8,6 +8,8 @@ import java.util.Map;
 import java.util.Set;
 
 import org.apache.logging.log4j.LogManager;
+
+import com.github.bartimaeusnek.ASM.CropStickTransformer_plugin;
 
 import cpw.mods.fml.common.DummyModContainer;
 import cpw.mods.fml.common.LoadController;
@@ -24,7 +26,6 @@ import cpw.mods.fml.common.eventhandler.EventBus;
 import cpw.mods.fml.common.versioning.ArtifactVersion;
 import cpw.mods.fml.common.versioning.InvalidVersionSpecificationException;
 import cpw.mods.fml.common.versioning.VersionRange;
-import lokko12.ASM.CropStickTransformer_plugin;
 
 @Mod(
 		modid = CropLoadCore.MODID, name = CropLoadCore.MODNAME, version = CropLoadCore.VERSION,
@@ -32,22 +33,15 @@ import lokko12.ASM.CropStickTransformer_plugin;
 		)
 
 
-public class CropLoadCore extends DummyModContainer {
+public class CropLoadCore {
 	
 	public static final String MODID = "croploadcore";
 	public static final String MODNAME = "CropLoadCore";
 	public static final String VERSION = "@version@";
 	public static final org.apache.logging.log4j.Logger CLClogger = LogManager.getLogger("CropLoadCore");
-	public static ModMetadata metadata = new ModMetadata();
+
 	@Instance(value = "CropLoadCore")
 	public static CropLoadCore instance;
-	
-	public CropLoadCore() {
-		metadata.modId = CropLoadCore.MODID;
-		metadata.name = CropLoadCore.MODNAME;
-		metadata.version = CropLoadCore.VERSION;
-		metadata.authorList.add("bartimaeusnek");
-	}
 	
 	@EventHandler
 	public void preInit(FMLPreInitializationEvent preinit) {
@@ -57,42 +51,5 @@ public class CropLoadCore extends DummyModContainer {
 	@EventHandler
 	public void Init(FMLInitializationEvent init) {
 	OreDict.register();
-	}
-
-	@Override
-	public String getModId() {
-		return MODID;
-	}
-
-	@Override
-	public String getName() {
-		return MODNAME;
-	}
-
-	@Override
-	public String getVersion() {
-		return VERSION;
-	}
-
-
-	@Override
-	public ModMetadata getMetadata() {
-		return metadata;
-	}
-
-	@Override
-	public boolean registerBus(com.google.common.eventbus.EventBus bus, LoadController controller) {
-		bus.register(this);
-		return true;
-	}
-
-	@Override
-	public boolean matches(Object mod) {
-		return instance.equals(mod);
-	}
-
-	@Override
-	public Object getMod() {
-		return instance;
 	}
 }
